Moved seedfiles to their own folder
